requirements.txt
Django~=4.2.4
django-jazzmin
reactpy-django
channels[daphne]
bootstrap4
django-phonenumber-field
python-decouple
whitenoise
django-cors-headers
django-floppyforms
django-widget-tweaks
django-crispy-forms==2.0
django-filter
gunicorn
drf-yasg
django-private-storage
pandas
django-admin-adminlte
django-xadmin
django-browser-reload
django-bootstrap-modal-forms
